Rename infection config file
Laminas matrix action now runs infection based on the presence of the file infection.json. Because it runs via phpdbg, it appears to crash due to running the mock server process. Renaming the config file should hopefully skip the matrix run and instead use the workflow that was already setup.